fix: address codex review round 1 findings
- Preserve org-level semantics when ancestorSources is empty by adding an
  explicit RenderInputs.ReadPlatformResources flag instead of inferring
  the render level from len(ancestorSources). Org/folder-level renders
  with zero ancestor templates now still read platformResources from the
  deployment template, matching pre-HOL-563 byte-identical semantics.
- Keep render.go at exactly one public render entry point (CueRenderer.Render)
  by moving the transitional raw-CUE helper (EvaluateGroupedCUE and
  evaluateCueInput) into render_raw_cue.go. The helper will be removed in
  Phase 3 once the templates package stops assembling raw CUE.
- Simplify the deployments test helpers: renderFlatWithAncestors and
  renderGroupedWithAncestors now set ReadPlatformResources=true
  explicitly, replacing the empty-ancestor placeholder workaround.
